abstraction for ls documentation
