+++ src/main/java/net/prominic/groovyls/compiler/ILanguageServerContext.java
@@ -2,6 +2,7 @@ package net.prominic.groovyls.compiler;
 
 import com.cleanroommc.groovyscript.sandbox.GroovySandbox;
 import io.github.classgraph.ScanResult;
+import net.prominic.groovyls.compiler.documentation.DocumentationFactory;
 import net.prominic.groovyls.util.FileContentsTracker;
 
 public interface ILanguageServerContext {
@@ -11,4 +12,6 @@ public interface ILanguageServerContext {
     ScanResult getScanResult();
 
     FileContentsTracker getFileContentsTracker();
+
+    DocumentationFactory getDocumentationFactory();
 }
